Added Getters
I added three functions. One returns the letter for the current node; another returns the letter for a node immediately left; and the last function returns a letter for the node immediately right. The function will return an empty char, if left or right is NULL. I also have a message that will display if the nodes are NULL, but if that causes problems, you can just comment those lines out (in Binary_Tree.cpp)